Style issues.
git-svn-id: http://argouml.tigris.org/svn/argouml/trunk/src_new@10021 a161b567-7d1e-0410-9ef9-912c70fedb3f

+++ org/argouml/uml/diagram/static_structure/layout/ClassdiagramInheritanceEdge.java
@@ -64,7 +64,7 @@ public abstract class ClassdiagramInheritanceEdge extends ClassdiagramEdge {
      * @return the vertical offset
      */
     public int getVerticalOffset() {
-        return (int) (getVGap() / 2) - 10 + getOffset();
+        return (getVGap() / 2) - 10 + getOffset();
     }
 
     /**
